% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketarg.R
\docType{package}
\name{basketarg-package}
\alias{basketarg}
\alias{basketarg-package}
\title{basketarg: Estadísticas de equipos y jugadores de la Liga Nacional de Básquet Argentina}
\description{
Este paquete contiene dos datasets que resumen el rendimiento de los equipos y
jugadores en partidos de la Liga Nacional de Básquet Argentina correspondentes a las
temporadas 2021-2022, 2022-2023 y 2023-2024. Los datos fueron obtenidos
mediante técnicas de web scraping desde un sitio web de estadísticas deportivas.
Incluye variables que describen la temporada, la condición, los resultados y el
desempeño general de cada equipo y jugador en cada partido.
}
\author{
\strong{Maintainer}: Nicolás Gallo \email{nicolasgallo98@hotmail.com}

}
\keyword{internal}
